Rename stackit errors package to stackiterrors
