+++ pkg/stackit/stackiterrors/errors.go
@@ -1,4 +1,4 @@
-package errors
+package stackiterrors
 
 import (
 	"errors"
