+++ pkg/stackit/stackiterrors/suite_test.go
@@ -1,4 +1,4 @@
-package errors
+package stackiterrors
 
 import (
 	"testing"
