+++ pkg/stackit/volumes.go
@@ -6,14 +6,13 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/stackitcloud/cloud-provider-stackit/pkg/stackit/stackiterrors"
 	"github.com/stackitcloud/stackit-sdk-go/core/runtime"
 	"github.com/stackitcloud/stackit-sdk-go/services/iaas"
 	sdkWait "github.com/stackitcloud/stackit-sdk-go/services/iaas/wait"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/klog/v2"
 	"k8s.io/utils/ptr"
-
-	stackiterrors "github.com/stackitcloud/cloud-provider-stackit/pkg/stackit/errors"
 )
 
 const (
